+++ commons/src/main/java/com/navercorp/pinpoint/common/trace/ServiceTypeProvider.java
@@ -25,7 +25,7 @@ public class ServiceTypeProvider {
 
     private static final ServiceTypeLocator UNREGISTERED = new ServiceTypeLocator() {
         @Override
-        public ServiceType findServiceType(short code) {
+        public ServiceType findServiceType(int code) {
             throw new IllegalStateException("ServiceTypeRegistry not registered");
         }
 
@@ -47,7 +47,7 @@ public class ServiceTypeProvider {
     }
 
     public static ServiceType getByCode(int serviceTypeCode) {
-        ServiceType serviceType = registry.findServiceType((short) serviceTypeCode);
+        ServiceType serviceType = registry.findServiceType(serviceTypeCode);
         if (ServiceType.UNDEFINED == serviceType) {
             throw new IllegalStateException("Unknown ServiceType code: " + serviceTypeCode);
         }
